requirements.txt
torch
torchvision
numpy
pillow
tqdm
datasets
huggingface_hubdiffusers
accelerate
transformers
